refactor: move job detail page to modules/job/detail
