+++ apps/web/src/modules/job/detail/components/JobCard/components/JobCardFooter/type.ts
