+++ apps/web/src/modules/job/detail/components/JobCard/index.tsx
