+++ apps/web/src/pages/job/[jid].tsx
@@ -1,3 +1,3 @@
-import JobDetailPage from 'modules/job/pages'
+import JobDetailPage from 'modules/job/detail/pages'
 
 export default JobDetailPage
